dedicated installer API
